+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Select/Action.js
