+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Select/DisplayValue.js
