+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Select/Divider.js
