+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Select/Option.js
