+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Select/action.scss
