+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Select/displayValue.scss
