+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Select/divider.scss
