+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Select/option.scss
