+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Select/select.scss
