+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Select/tests/Action.test.js
