+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Select/tests/DisplayValue.test.js
