+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Select/tests/Divider.test.js
