+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Select/tests/Option.test.js
