+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Select/types.js
